Bump jest and ts-jest to be consistent across backend packages to fix CI issue, test config updates
