+++ packages/backend-core/jest.config.ts
@@ -3,7 +3,7 @@ import { Config } from "@jest/types"
 const config: Config.InitialOptions = {
   preset: "ts-jest",
   testEnvironment: "node",
-  setupFiles: ["./scripts/jestSetup.ts"],
+  setupFiles: ["./tests/jestSetup.ts"],
   collectCoverageFrom: ["src/**/*.{js,ts}"],
   coverageReporters: ["lcov", "json", "clover"],
 }
