+++ packages/backend-core/package.json
@@ -67,12 +67,12 @@
     "@types/uuid": "8.3.4",
     "chance": "1.1.3",
     "ioredis-mock": "5.8.0",
-    "jest": "27.5.1",
+    "jest": "28.1.1",
     "koa": "2.7.0",
     "nodemon": "2.0.16",
     "pouchdb-adapter-memory": "7.2.2",
     "timekeeper": "2.2.0",
-    "ts-jest": "27.1.5",
+    "ts-jest": "28.0.4",
     "ts-node": "10.8.1",
     "tsconfig-paths": "4.0.0",
     "typescript": "4.7.3"
