+++ packages/backend-core/src/db/tests/index.spec.js
@@ -1,4 +1,4 @@
-require("../../../tests/utilities/TestConfiguration")
+require("../../../tests")
 const { dangerousGetDB } = require("../")
 
 describe("db", () => {
