+++ packages/backend-core/src/middleware/passport/tests/third-party-common.spec.js
@@ -1,4 +1,4 @@
-require("../../../../tests/utilities/TestConfiguration")
+require("../../../../tests")
 const { authenticateThirdParty } = require("../third-party-common")
 const { data } = require("./utilities/mock-data")
 const { DEFAULT_TENANT_ID } = require("../../../constants")
