--- packages/backend-core/tests/utilities/TestConfiguration.js
@@ -1,1 +1,0 @@
-require("./db")
