--- packages/backend-core/tests/utilities/db.js
@@ -1,6 +1,0 @@
-const core = require("../../src/index")
-const dbConfig = {
-  inMemory: true,
-  allDbs: true,
-}
-core.init({ db: dbConfig })
