+++ packages/backend-core/tests/utilities/db.ts
@@ -1,0 +1,9 @@
+import * as db from "../../src/db"
+
+const dbConfig = {
+  inMemory: true,
+}
+
+export const init = () => {
+  db.init(dbConfig)
+}
