+++ packages/backend-core/tests/utilities/index.ts
@@ -1,2 +1,5 @@
 export * as mocks from "./mocks"
 export * as structures from "./structures"
+
+import * as dbConfig from "./db"
+dbConfig.init()
