+++ packages/builder/package.json
@@ -117,7 +117,7 @@
     "start-server-and-test": "^1.12.1",
     "svelte": "^3.48.0",
     "svelte-jester": "^1.3.2",
-    "ts-node": "10.8.2",
+    "ts-node": "10.8.1",
     "tsconfig-paths": "4.0.0",
     "typescript": "4.7.3",
     "vite": "^3.0.8"
