+++ packages/server/package.json
@@ -145,7 +145,7 @@
     "eslint": "6.8.0",
     "ioredis-mock": "7.2.0",
     "is-wsl": "2.2.0",
-    "jest": "27.5.1",
+    "jest": "28.1.1",
     "jest-openapi": "0.14.2",
     "nodemon": "2.0.15",
     "openapi-types": "9.3.1",
@@ -156,7 +156,7 @@
     "supertest": "4.0.2",
     "swagger-jsdoc": "6.1.0",
     "timekeeper": "2.2.0",
-    "ts-jest": "27.1.3",
+    "ts-jest": "28.0.4",
     "ts-node": "10.8.1",
     "tsconfig-paths": "4.0.0",
     "typescript": "4.7.3",
