+++ packages/server/src/tests/jestSetup.ts
@@ -1,5 +1,5 @@
 const { tmpdir } = require("os")
-const env = require("../src/environment")
+import env from "../environment"
 
 env._set("SELF_HOSTED", "1")
 env._set("NODE_ENV", "jest")
@@ -9,7 +9,7 @@ env._set("BUDIBASE_DIR", tmpdir("budibase-unittests"))
 env._set("LOG_LEVEL", "silent")
 env._set("PORT", 0)
 
-const { mocks } = require("@budibase/backend-core/tests")
+import { mocks } from "@budibase/backend-core/tests"
 
 // mock all dates to 2020-01-01T00:00:00.000Z
 // use tk.reset() to use real dates in individual tests
