+++ packages/worker/jest.config.ts
@@ -4,7 +4,7 @@ import * as fs from "fs"
 const config: Config.InitialOptions = {
   preset: "ts-jest",
   testEnvironment: "node",
-  setupFiles: ["./scripts/jestSetup.ts"],
+  setupFiles: ["./src/tests/jestSetup.ts"],
   collectCoverageFrom: ["src/**/*.{js,ts}"],
   coverageReporters: ["lcov", "json", "clover"],
 }
