+++ packages/worker/src/index.ts
@@ -13,7 +13,7 @@ import { Scope } from "@sentry/node"
 import { Event } from "@sentry/types/dist/event"
 import Application from "koa"
 import { bootstrap } from "global-agent"
-import db from "./db"
+import * as db from "./db"
 db.init()
 const Koa = require("koa")
 const destroyable = require("server-destroy")
